+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMG5HC4Q2SCS388V6STREA1C"
+        "01KMHQ6ETCC171TFD67ZHCWS3G"
     )
     r"""The unique identifier of the data source"""